move common models and validation functions to common file
